Fix issue #10 'Bug in the Lexer'
Introduced a property of a TokenType to be a delimiter.
Modified regexp so that every non-delimiter token is wrapped by
delimiter tokens (or start/end of a string) at both sides.
Used 0-length lookahead and lookbehind to implement this.
Also modified INVALID_TOKEN regexp so that it captures all symbols of
unrecognized pattern and stops when it sees a delimiter type of token.
This makes it easier to describe a lexical error in the source.