Updates to Lexer's error detection system
Add non-greedy quantifier to INVALID_TOKEN TokenType, so that it
doesn't report all the line as an error, but only reports the token up
to a first delimiter character in the line.
Introduce a custom exception for lexical errors: LexicalError.java.
It now reports a line and a column in the text where the invalid token
is located and also reports the invalid token itself.