Redesign a project structure and a build system
Now /src/test/java is marked as a test folder and test classes will be
compiled into /test/ directory.
Introduced ant build-file build.xml so that the project can be built
with Ant. The build-file can compile and run the compiled code. It can
also compile and run tests and produce a tests report in html format.
And it can also build a jar file of the whole application.
Also changed Eclipse launchers file so that we can run WhileINterpreter
class and eclipse will provide arguments to it.