Restructure AST classes, modify Assign Cmd Record
All AST related classes now live in a single package.
Assign record now takes an instance of Variable record as an argument
corresponding to the variable to be assigned to (instead of a String
representing the variable name).
This eliminates the need for validating the correctness of the variable
name in the Assign, as it is already handled by the Variable record.
Now the Variable record doesn't only represent a construct specific to
ArithmeticExpression, but rather being its own entity, which can also be
used inside arithmetic expression (hence the reason to restructure
whileinterpreter.ast package).