Outline basic architecture of the GUI application
Try out a kind of MVCI pattern:

Model (ViewModel) is a collection of JavaFX properties representing
the state of the gui (but need not mirror the views exactly).

Views are fxml pages every user action in the view will be redirected to
the corresponding controller.

Controllers are components that glue all the other components together.
They have 2 responsibilities:
1. Bind the (fxml) Views to the ViewModel (initialization of controller).
2. Convert user actions in the View to the calls to the Interactors.

Interactors are Actions, each Interactor corresponds to single action
(e.g. load file, execute program, and others). Interactors only depend
on the ViewModel, making them decoupled from JavaFX views.
This has 2 benefits:
1. Easier testability - we do not need to launch a JavaFX application
when we want to unit-test an Action making the tests easier to write
and faster to run. However, we still need integration tests to make sure
we correctly glued fxml Views to the ViewModel in the Controllers.
2. The same Action can be invoked from multiple Views without any code
duplication and without coupling between the views themselves.